Restructuring: contact form clean up
