--- src/components/ui/Label.tsx
@@ -1,5 +1,0 @@
-export default function Label() {
-  return (
-    <div>Label</div>
-  );
-}
